+++ packages/charts/src/components/PieChart/PieChart.stories.tsx
@@ -1,6 +1,6 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { type Meta, type StoryFn } from '@storybook/react';
+import { type Meta, type StoryFn } from '@storybook/react-vite';
 
 import { PieChart, type PieChartDatum, type PieChartProps } from './';