+++ packages/components/eslint.config.mjs
@@ -4,13 +4,13 @@
 
 // @ts-expect-error No types available
 import frontifyConfig from '@frontify/eslint-config-react';
+import { defineConfig } from 'eslint/config';
 // @ts-expect-error No types available
 import noticePlugin from 'eslint-plugin-notice';
 // @ts-expect-error No types available
 import tailwindcssPlugin from 'eslint-plugin-tailwindcss';
-import tseslint from 'typescript-eslint';
 
-export default tseslint.config(
+export default defineConfig(
     {
         ignores: [
             'dist/',