+++ packages/components/src/components/Flyout/Flyout.tsx
@@ -140,6 +140,7 @@ export const FlyoutContent = (
         rounded = 'medium',
         width = 'fit-content',
         shadow = 'medium',
+        side,
         triggerOffset = 'compact',
         viewportCollisionPadding = 'compact',
         'data-test-id': dataTestId = 'fondue-flyout-content',
@@ -148,10 +149,26 @@ export const FlyoutContent = (
     }: FlyoutContentProps,
     ref: ForwardedRef<HTMLDivElement>,
 ) => {
-    const theme = useFondueTheme();
+    const { theme, dir } = useFondueTheme();
+
+    const getAdjustedSide = (side?: 'top' | 'right' | 'bottom' | 'left') => {
+        if (!side || dir === 'ltr') {
+            return side;
+        }
+
+        if (side === 'left') {
+            return 'right';
+        }
+        if (side === 'right') {
+            return 'left';
+        }
+
+        return side;
+    };
+
     return (
         <RadixPopover.Portal>
-            <ThemeProvider theme={theme}>
+            <ThemeProvider theme={theme} dir={dir}>
                 <div data-test-id="fondue-flyout-overlay" className={styles.overlay} />
                 <RadixPopover.Content
                     style={
@@ -164,6 +181,7 @@ export const FlyoutContent = (
                     align={align}
                     collisionPadding={VIEWPORT_COLLISION_PADDING_MAP[viewportCollisionPadding]}
                     sideOffset={SPACING_MAP[triggerOffset]}
+                    side={getAdjustedSide(side)}
                     className={styles.root}
                     data-flyout-spacing={padding}
                     data-rounded={rounded}