+++ packages/fondue/src/components/DatePicker/DatePickerTrigger.tsx
@@ -18,11 +18,23 @@ type DatePickerTriggerProps = {
     onClick?: () => void;
     validation?: Validation;
     onDateChanged?: ((date: [Date | null, Date | null] | null) => void) | ((date: Date | null) => void);
+    hugWidth?: boolean;
+    'aria-label': string;
 };
 
 export const DatePickerTrigger = forwardRef<HTMLDivElement, DatePickerTriggerProps>(
     (
-        { value, onClick, isClearable, placeHolder, isCalendarOpen, validation = Validation.Default, onDateChanged },
+        {
+            value,
+            onClick,
+            isClearable,
+            placeHolder,
+            isCalendarOpen,
+            validation = Validation.Default,
+            onDateChanged,
+            hugWidth = true,
+            'aria-label': ariaLabel,
+        },
         ref,
     ) => {
         const isWarningOrErrorValidationState = validation === Validation.Error || validation === Validation.Warning;
@@ -51,7 +63,7 @@ export const DatePickerTrigger = forwardRef<HTMLDivElement, DatePickerTriggerPro
         );
 
         return (
-            <div ref={ref}>
+            <div ref={ref} className={hugWidth ? 'tw-w-auto' : 'tw-w-full'}>
                 <div
                     className={merge([
                         'tw-absolute tw-top-2 tw-text-black-60 tw-z-[1]',
@@ -73,6 +85,7 @@ export const DatePickerTrigger = forwardRef<HTMLDivElement, DatePickerTriggerPro
                             onClick?.();
                         }
                     }}
+                    aria-label={ariaLabel}
                     data-test-id="open-close-click-zone"
                 ></div>
                 <TextInput