+++ packages/fondue/src/components/EditableText/EditableText.stories.tsx
@@ -1,6 +1,6 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { type StoryFn } from '@storybook/react';
+import { type StoryFn } from '@storybook/react-vite';
 import { useState } from 'react';
 
 import { EditableMode, EditableText, type EditableTextProps } from '@components/EditableText/EditableText';