+++ packages/fondue/src/components/Table/Table.stories.tsx
@@ -1,8 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { action } from '@storybook/addon-actions';
-import { type Meta, type StoryFn } from '@storybook/react';
+import { type Meta, type StoryFn } from '@storybook/react-vite';
 import { type ReactElement, useEffect, useState } from 'react';
+import { action } from 'storybook/actions';
 
 import { Badge } from '@components/Badge/Badge';
 import { Button } from '@components/Button/Button';
@@ -23,6 +23,11 @@ export default {
     tags: ['autodocs'],
     args: {
         selectionMode: SelectionMode.NoSelect,
+    },
+    parameters: {
+        status: {
+            type: 'deprecated',
+        },
     },
     argTypes: {
         type: {