+++ packages/tokens/eslint.config.mjs
@@ -2,12 +2,15 @@
 
 // @ts-check
 
+// @ts-expect-error No types available
 import frontifyConfig from '@frontify/eslint-config-react';
+import { defineConfig } from 'eslint/config';
+// @ts-expect-error No types available
 import noticePlugin from 'eslint-plugin-notice';
+// @ts-expect-error No types available
 import tailwindcssPlugin from 'eslint-plugin-tailwindcss';
-import tseslint from 'typescript-eslint';
 
-export default tseslint.config(
+export default defineConfig(
     {
         ignores: ['tmp/', 'dist/', 'node_modules/'],
     },